test: remove provider related test from callback registry test suite
