+++ vda5050_execution/test/unit/test_callback_registry.cpp
@@ -22,7 +22,6 @@
 
 #include "vda5050_execution/callback_registry.hpp"
 #include "vda5050_execution/event.hpp"
-#include "vda5050_execution/update.hpp"
 
 struct EventA : public vda5050_execution::EventBase
 {
@@ -37,22 +36,6 @@ struct EventB : public vda5050_execution::EventBase
   std::type_index get_type() const override
   {
     return typeid(EventB);
-  }
-};
-
-struct UpdateA : public vda5050_execution::UpdateBase
-{
-  std::type_index get_type() const override
-  {
-    return typeid(UpdateA);
-  }
-};
-
-struct UpdateB : public vda5050_execution::UpdateBase
-{
-  std::type_index get_type() const override
-  {
-    return typeid(UpdateB);
   }
 };
 
@@ -103,56 +86,10 @@ TEST(CallbackRegistryTest, MultiEventDispatch)
   EXPECT_EQ(call_count_2, 1);
 }
 
-TEST(CallbackRegistryTest, SingleUpdateQuery)
-{
-  vda5050_execution::CallbackRegistry registry;
-
-  int call_count = 0;
-  int wront_type_count = 0;
-
-  registry.register_provider<UpdateA>(
-    [&](std::shared_ptr<UpdateA>) { call_count++; });
-  registry.register_provider<UpdateB>(
-    [&](std::shared_ptr<UpdateB>) { wront_type_count++; });
-
-  auto update = std::make_shared<UpdateA>();
-  registry.query(update);
-
-  EXPECT_EQ(call_count, 1);
-  EXPECT_EQ(wront_type_count, 0);
-}
-
-TEST(CallbackRegistryTest, MultiUpdateQuery)
-{
-  vda5050_execution::CallbackRegistry registry;
-
-  int call_count_1 = 0;
-  int call_count_2 = 0;
-
-  registry.register_provider<UpdateA>(
-    [&](std::shared_ptr<UpdateA>) { call_count_1++; });
-  registry.register_provider<UpdateB>(
-    [&](std::shared_ptr<UpdateB>) { call_count_2++; });
-
-  auto update_a = std::make_shared<UpdateA>();
-  registry.query(update_a);
-
-  auto update_b = std::make_shared<UpdateB>();
-  registry.query(update_b);
-
-  registry.query(update_a);
-
-  EXPECT_EQ(call_count_1, 2);
-  EXPECT_EQ(call_count_2, 1);
-}
-
 TEST(CallbackRegistryTest, EmptyCalls)
 {
   vda5050_execution::CallbackRegistry registry;
 
   auto event = std::make_shared<EventA>();
   EXPECT_NO_THROW(registry.dispatch(event));
-
-  auto update = std::make_shared<UpdateA>();
-  EXPECT_NO_THROW(registry.query(update));
 }
