+++ vda5050_execution/test/unit/test_provider.cpp
@@ -24,6 +24,69 @@
 #include "vda5050_execution/provider.hpp"
 #include "vda5050_execution/update.hpp"
 
+struct UpdateA : public vda5050_execution::UpdateBase
+{
+  std::type_index get_type() const override
+  {
+    return typeid(UpdateA);
+  }
+};
+
+struct UpdateB : public vda5050_execution::UpdateBase
+{
+  std::type_index get_type() const override
+  {
+    return typeid(UpdateB);
+  }
+};
+
+TEST(ProviderRegistryTest, SingleUpdateQuery)
+{
+  vda5050_execution::Provider provider;
+
+  int call_count = 0;
+  int wront_type_count = 0;
+
+  provider.on<UpdateA>([&](std::shared_ptr<UpdateA>) { call_count++; });
+  provider.on<UpdateB>([&](std::shared_ptr<UpdateB>) { wront_type_count++; });
+
+  auto update = std::make_shared<UpdateA>();
+  provider.push_shared(update);
+
+  EXPECT_EQ(call_count, 1);
+  EXPECT_EQ(wront_type_count, 0);
+}
+
+TEST(ProviderRegistryTest, MultiUpdateQuery)
+{
+  vda5050_execution::Provider provider;
+
+  int call_count_1 = 0;
+  int call_count_2 = 0;
+
+  provider.on<UpdateA>([&](std::shared_ptr<UpdateA>) { call_count_1++; });
+  provider.on<UpdateB>([&](std::shared_ptr<UpdateB>) { call_count_2++; });
+
+  auto update_a = std::make_shared<UpdateA>();
+  provider.push_shared(update_a);
+
+  auto update_b = std::make_shared<UpdateB>();
+  provider.push_shared(update_b);
+
+  provider.push_shared(update_a);
+
+  EXPECT_EQ(call_count_1, 2);
+  EXPECT_EQ(call_count_2, 1);
+}
+
+TEST(ProviderRegistryTest, EmptyCalls)
+{
+  vda5050_execution::Provider provider;
+
+  auto update = std::make_shared<UpdateA>();
+  EXPECT_NO_THROW(provider.push_shared(update));
+}
+
 TEST(ProviderTest, PushUpdate)
 {
   auto provider = std::make_shared<vda5050_execution::Provider>();
